pyproject.toml
[project]
name = "open-source-python-template"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
  { name = "ospsd", email = "wz1509@nyu.edu" },
]
dependencies = [
    "flask>=3.0.3",
    "psycopg2-binary>=2.9.9",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.127.0",
    "selenium>=4.20.0",
]

requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "flake8>=7.0.0",
    "black>=24.1.1",
    "types-psycopg2>=2.9.21.20240417",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
